# front_end/openapi(3).json
{"openapi":"3.1.0","info":{"title":"FastAPI","description":"University Portal API.","version":"0.1.0"},"paths":{"/":{"get":{"summary":"Index","operationId":"index__get","responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{}}}}}}},"/auth/token":{"post":{"summary":"Login","operationId":"login_auth_token_post","requestBody":{"content":{"application/x-www-form-urlencoded":{"schema":{"$ref":"#/components/schemas/Body_login_auth_token_post"}}},"required":true},"responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{"$ref":"#/components/schemas/Token"}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}}},"/faces/auth/register":{"post":{"summary":"Register Face","operationId":"register_face_faces_auth_register_post","security":[{"OAuth2PasswordBearer":[]}],"parameters":[{"name":"image_data","in":"query","required":true,"schema":{"type":"string","title":"Image Data"}},{"name":"user_id","in":"query","required":true,"schema":{"type":"string","format":"uuid","title":"User Id"}}],"responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}}},"/faces/auth/login":{"post":{"summary":"Login With Face","operationId":"login_with_face_faces_auth_login_post","parameters":[{"name":"image_data","in":"query","required":true,"schema":{"type":"string","title":"Image Data"}}],"responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}}},"/faces/me":{"get":{"summary":"Get My Faces","operationId":"get_my_faces_faces_me_get","security":[{"OAuth2PasswordBearer":[]}],"parameters":[{"name":"user_id","in":"query","required":true,"schema":{"type":"string","format":"uuid","title":"User Id"}}],"responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{"type":"array","items":{"$ref":"#/components/schemas/FaceEncodingOut"},"title":"Response Get My Faces Faces Me Get"}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}}},"/faces/{face_id}":{"delete":{"summary":"Delete Face","operationId":"delete_face_faces__face_id__delete","security":[{"OAuth2PasswordBearer":[]}],"parameters":[{"name":"face_id","in":"path","required":true,"schema":{"type":"string","format":"uuid","title":"Face Id"}}],"responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}}},"/users/":{"post":{"summary":"Create New User","operationId":"create_new_user_users__post","requestBody":{"required":true,"content":{"application/json":{"schema":{"$ref":"#/components/schemas/UserIn"}}}},"responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{"$ref":"#/components/schemas/Msg"}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}},"get":{"summary":"Get Users","operationId":"get_users_users__get","security":[{"OAuth2PasswordBearer":[]}],"parameters":[{"name":"skip","in":"query","required":false,"schema":{"type":"integer","minimum":0,"default":0,"title":"Skip"}},{"name":"limit","in":"query","required":false,"schema":{"type":"integer","minimum":1,"default":100,"title":"Limit"}}],"responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{"type":"array","items":{"$ref":"#/components/schemas/UserOut"},"title":"Response Get Users Users  Get"}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}}},"/users/me":{"get":{"summary":"Get User Me","operationId":"get_user_me_users_me_get","responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{"$ref":"#/components/schemas/UserOut"}}}}},"security":[{"OAuth2PasswordBearer":[]}]}},"/users/{user_id}":{"get":{"summary":"Get User By Id","operationId":"get_user_by_id_users__user_id__get","security":[{"OAuth2PasswordBearer":[]}],"parameters":[{"name":"user_id","in":"path","required":true,"schema":{"type":"string","format":"uuid","title":"User Id"}}],"responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{"$ref":"#/components/schemas/UserOut"}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}},"patch":{"summary":"Patch User","operationId":"patch_user_users__user_id__patch","security":[{"OAuth2PasswordBearer":[]}],"parameters":[{"name":"user_id","in":"path","required":true,"schema":{"type":"string","format":"uuid","title":"User Id"}}],"requestBody":{"required":true,"content":{"application/json":{"schema":{"$ref":"#/components/schemas/UserUpdate"}}}},"responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{"$ref":"#/components/schemas/UserOut"}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}},"delete":{"summary":"Delete User By Id","operationId":"delete_user_by_id_users__user_id__delete","security":[{"OAuth2PasswordBearer":[]}],"parameters":[{"name":"user_id","in":"path","required":true,"schema":{"type":"string","format":"uuid","title":"User Id"}}],"responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{"$ref":"#/components/schemas/Msg"}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}}},"/users/{user_id}/role":{"patch":{"summary":"Patch User By Admin","operationId":"patch_user_by_admin_users__user_id__role_patch","security":[{"OAuth2PasswordBearer":[]}],"parameters":[{"name":"user_id","in":"path","required":true,"schema":{"type":"string","format":"uuid","title":"User Id"}}],"requestBody":{"required":true,"content":{"application/json":{"schema":{"$ref":"#/components/schemas/UserAdminUpdate"}}}},"responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{"$ref":"#/components/schemas/UserOut"}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}}},"/courses/":{"post":{"summary":"Create New Course","operationId":"create_new_course_courses__post","security":[{"OAuth2PasswordBearer":[]}],"requestBody":{"required":true,"content":{"application/json":{"schema":{"$ref":"#/components/schemas/CourseIn"}}}},"responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{"$ref":"#/components/schemas/Msg"}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}},"get":{"summary":"Get Courses","operationId":"get_courses_courses__get","security":[{"OAuth2PasswordBearer":[]}],"parameters":[{"name":"skip","in":"query","required":false,"schema":{"type":"integer","minimum":0,"default":0,"title":"Skip"}},{"name":"limit","in":"query","required":false,"schema":{"type":"integer","minimum":1,"default":100,"title":"Limit"}}],"responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{"type":"array","items":{"$ref":"#/components/schemas/CourseOut"},"title":"Response Get Courses Courses  Get"}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}}},"/courses/{course_id}":{"get":{"summary":"Get Course By Id","operationId":"get_course_by_id_courses__course_id__get","security":[{"OAuth2PasswordBearer":[]}],"parameters":[{"name":"course_id","in":"path","required":true,"schema":{"type":"string","format":"uuid","title":"Course Id"}}],"responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{"$ref":"#/components/schemas/CourseOut"}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}},"patch":{"summary":"Patch Course","operationId":"patch_course_courses__course_id__patch","security":[{"OAuth2PasswordBearer":[]}],"parameters":[{"name":"course_id","in":"path","required":true,"schema":{"type":"string","format":"uuid","title":"Course Id"}}],"requestBody":{"required":true,"content":{"application/json":{"schema":{"$ref":"#/components/schemas/CourseUpdate"}}}},"responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{"$ref":"#/components/schemas/CourseOut"}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}},"delete":{"summary":"Delete Course By Id","operationId":"delete_course_by_id_courses__course_id__delete","security":[{"OAuth2PasswordBearer":[]}],"parameters":[{"name":"course_id","in":"path","required":true,"schema":{"type":"string","format":"uuid","title":"Course Id"}}],"responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{"$ref":"#/components/schemas/Msg"}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}}},"/faculties/":{"post":{"summary":"Create New Faculty","operationId":"create_new_faculty_faculties__post","security":[{"OAuth2PasswordBearer":[]}],"requestBody":{"required":true,"content":{"application/json":{"schema":{"$ref":"#/components/schemas/FacultyIn"}}}},"responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{"$ref":"#/components/schemas/Msg"}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}},"get":{"summary":"Get Faculties","operationId":"get_faculties_faculties__get","security":[{"OAuth2PasswordBearer":[]}],"parameters":[{"name":"skip","in":"query","required":false,"schema":{"type":"integer","minimum":0,"default":0,"title":"Skip"}},{"name":"limit","in":"query","required":false,"schema":{"type":"integer","minimum":1,"default":100,"title":"Limit"}}],"responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{"type":"array","items":{"$ref":"#/components/schemas/FacultyOutWithDepartment"},"title":"Response Get Faculties Faculties  Get"}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}}},"/faculties/{faculty_code}":{"get":{"summary":"Get Faculty By Id","operationId":"get_faculty_by_id_faculties__faculty_code__get","security":[{"OAuth2PasswordBearer":[]}],"parameters":[{"name":"faculty_code","in":"path","required":true,"schema":{"type":"string","title":"Faculty Code"}}],"responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{"$ref":"#/components/schemas/FacultyOutWithDepartment"}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}},"patch":{"summary":"Patch Faculty","operationId":"patch_faculty_faculties__faculty_code__patch","security":[{"OAuth2PasswordBearer":[]}],"parameters":[{"name":"faculty_code","in":"path","required":true,"schema":{"type":"string","title":"Faculty Code"}}],"requestBody":{"required":true,"content":{"application/json":{"schema":{"$ref":"#/components/schemas/FacultyUpdate"}}}},"responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{"$ref":"#/components/schemas/FacultyOut"}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}},"delete":{"summary":"Delete Faculty By Id","operationId":"delete_faculty_by_id_faculties__faculty_code__delete","security":[{"OAuth2PasswordBearer":[]}],"parameters":[{"name":"faculty_code","in":"path","required":true,"schema":{"type":"string","title":"Faculty Code"}}],"responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{"$ref":"#/components/schemas/Msg"}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}}},"/departments/":{"post":{"summary":"Create New Department","operationId":"create_new_department_departments__post","security":[{"OAuth2PasswordBearer":[]}],"requestBody":{"required":true,"content":{"application/json":{"schema":{"$ref":"#/components/schemas/DepartmentIn"}}}},"responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{"$ref":"#/components/schemas/Msg"}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}},"get":{"summary":"Get Departments","operationId":"get_departments_departments__get","security":[{"OAuth2PasswordBearer":[]}],"parameters":[{"name":"skip","in":"query","required":false,"schema":{"type":"integer","minimum":0,"default":0,"title":"Skip"}},{"name":"limit","in":"query","required":false,"schema":{"type":"integer","minimum":1,"default":100,"title":"Limit"}}],"responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{"type":"array","items":{"$ref":"#/components/schemas/DepartmentOut"},"title":"Response Get Departments Departments  Get"}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}}},"/departments/{department_id}":{"get":{"summary":"Get Department By Id","operationId":"get_department_by_id_departments__department_id__get","security":[{"OAuth2PasswordBearer":[]}],"parameters":[{"name":"department_id","in":"path","required":true,"schema":{"type":"string","format":"uuid","title":"Department Id"}}],"responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{"$ref":"#/components/schemas/DepartmentOut"}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}},"patch":{"summary":"Patch Department","operationId":"patch_department_departments__department_id__patch","security":[{"OAuth2PasswordBearer":[]}],"parameters":[{"name":"department_id","in":"path","required":true,"schema":{"type":"string","format":"uuid","title":"Department Id"}}],"requestBody":{"required":true,"content":{"application/json":{"schema":{"$ref":"#/components/schemas/DepartmentUpdate"}}}},"responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{"$ref":"#/components/schemas/DepartmentOut"}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}},"delete":{"summary":"Delete Department By Id","operationId":"delete_department_by_id_departments__department_id__delete","security":[{"OAuth2PasswordBearer":[]}],"parameters":[{"name":"department_id","in":"path","required":true,"schema":{"type":"string","format":"uuid","title":"Department Id"}}],"responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{"$ref":"#/components/schemas/Msg"}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}}},"/enrollments/":{"post":{"summary":"Enroll In Course","operationId":"enroll_in_course_enrollments__post","security":[{"OAuth2PasswordBearer":[]}],"requestBody":{"required":true,"content":{"application/json":{"schema":{"$ref":"#/components/schemas/Body_enroll_in_course_enrollments__post"}}}},"responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{"$ref":"#/components/schemas/Msg"}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}},"get":{"summary":"Get Enrollment","operationId":"get_enrollment_enrollments__get","security":[{"OAuth2PasswordBearer":[]}],"parameters":[{"name":"skip","in":"query","required":false,"schema":{"type":"integer","minimum":0,"default":0,"title":"Skip"}},{"name":"limit","in":"query","required":false,"schema":{"type":"integer","minimum":1,"default":100,"title":"Limit"}}],"requestBody":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/Body_get_enrollment_enrollments__get"}}}},"responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{"type":"array","items":{"$ref":"#/components/schemas/EnrollmentOut"},"title":"Response Get Enrollment Enrollments  Get"}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}}},"/enrollments/{enrollment_id}":{"get":{"summary":"Get Enrollment By Id","operationId":"get_enrollment_by_id_enrollments__enrollment_id__get","security":[{"OAuth2PasswordBearer":[]}],"parameters":[{"name":"enrollment_id","in":"path","required":true,"schema":{"type":"string","format":"uuid","title":"Enrollment Id"}}],"responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{"$ref":"#/components/schemas/EnrollmentOut"}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}},"delete":{"summary":"Delete Enrollment By Id","operationId":"delete_enrollment_by_id_enrollments__enrollment_id__delete","security":[{"OAuth2PasswordBearer":[]}],"parameters":[{"name":"enrollment_id","in":"path","required":true,"schema":{"type":"string","format":"uuid","title":"Enrollment Id"}}],"responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{"$ref":"#/components/schemas/Msg"}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}}}},"components":{"schemas":{"Body_enroll_in_course_enrollments__post":{"properties":{"course_id":{"type":"string","format":"uuid","title":"Course Id"}},"type":"object","required":["course_id"],"title":"Body_enroll_in_course_enrollments__post"},"Body_get_enrollment_enrollments__get":{"properties":{"course_id":{"anyOf":[{"type":"string","format":"uuid"},{"type":"null"}],"title":"Course Id"}},"type":"object","title":"Body_get_enrollment_enrollments__get"},"Body_login_auth_token_post":{"properties":{"grant_type":{"anyOf":[{"type":"string","pattern":"^password$"},{"type":"null"}],"title":"Grant Type"},"username":{"type":"string","title":"Username"},"password":{"type":"string","format":"password","title":"Password"},"scope":{"type":"string","title":"Scope","default":""},"client_id":{"anyOf":[{"type":"string"},{"type":"null"}],"title":"Client Id"},"client_secret":{"anyOf":[{"type":"string"},{"type":"null"}],"format":"password","title":"Client Secret"}},"type":"object","required":["username","password"],"title":"Body_login_auth_token_post"},"Course":{"properties":{"code":{"type":"string","title":"Code"},"title":{"type":"string","title":"Title"},"description":{"anyOf":[{"type":"string"},{"type":"null"}],"title":"Description"},"credits":{"type":"integer","title":"Credits","default":3},"level":{"type":"string","title":"Level"},"semester":{"type":"string","title":"Semester"},"department_id":{"type":"string","format":"uuid","title":"Department Id"},"id":{"anyOf":[{"type":"string","format":"uuid"},{"type":"null"}],"title":"Id"},"created_at":{"type":"string","format":"date-time","title":"Created At"}},"type":"object","required":["code","title","level","semester","department_id"],"title":"Course"},"CourseIn":{"properties":{"code":{"type":"string","title":"Code"},"title":{"type":"string","title":"Title"},"description":{"anyOf":[{"type":"string"},{"type":"null"}],"title":"Description"},"credits":{"type":"integer","title":"Credits","default":3},"level":{"type":"string","title":"Level"},"semester":{"type":"string","title":"Semester"},"department_id":{"type":"string","format":"uuid","title":"Department Id"}},"type":"object","required":["code","title","level","semester","department_id"],"title":"CourseIn"},"CourseOut":{"properties":{"code":{"type":"string","title":"Code"},"title":{"type":"string","title":"Title"},"description":{"anyOf":[{"type":"string"},{"type":"null"}],"title":"Description"},"credits":{"type":"integer","title":"Credits","default":3},"level":{"type":"string","title":"Level"},"semester":{"type":"string","title":"Semester"},"department_id":{"type":"string","format":"uuid","title":"Department Id"},"id":{"type":"string","format":"uuid","title":"Id"},"enrollments":{"items":{"$ref":"#/components/schemas/Enrollment"},"type":"array","title":"Enrollments"}},"type":"object","required":["code","title","level","semester","department_id","id","enrollments"],"title":"CourseOut"},"CourseUpdate":{"properties":{"code":{"anyOf":[{"type":"string"},{"type":"null"}],"title":"Code"},"title":{"anyOf":[{"type":"string"},{"type":"null"}],"title":"Title"},"description":{"anyOf":[{"type":"string"},{"type":"null"}],"title":"Description"},"credits":{"anyOf":[{"type":"integer","minimum":0.0},{"type":"null"}],"title":"Credits"},"level":{"anyOf":[{"type":"string"},{"type":"null"}],"title":"Level"},"semester":{"anyOf":[{"type":"string"},{"type":"null"}],"title":"Semester"},"department":{"anyOf":[{"type":"string"},{"type":"null"}],"title":"Department"}},"type":"object","title":"CourseUpdate"},"Department":{"properties":{"name":{"type":"string","title":"Name"},"description":{"anyOf":[{"type":"string"},{"type":"null"}],"title":"Description"},"faculty_id":{"type":"string","format":"uuid","title":"Faculty Id"},"id":{"anyOf":[{"type":"string","format":"uuid"},{"type":"null"}],"title":"Id"},"created_at":{"type":"string","format":"date-time","title":"Created At"}},"type":"object","required":["name","faculty_id"],"title":"Department"},"DepartmentIn":{"properties":{"name":{"type":"string","title":"Name"},"description":{"anyOf":[{"type":"string"},{"type":"null"}],"title":"Description"},"faculty_id":{"type":"string","format":"uuid","title":"Faculty Id"}},"type":"object","required":["name","faculty_id"],"title":"DepartmentIn"},"DepartmentOut":{"properties":{"name":{"type":"string","title":"Name"},"description":{"anyOf":[{"type":"string"},{"type":"null"}],"title":"Description"},"faculty_id":{"type":"string","format":"uuid","title":"Faculty Id"},"id":{"type":"string","format":"uuid","title":"Id"},"members":{"anyOf":[{"items":{"$ref":"#/components/schemas/User"},"type":"array"},{"type":"null"}],"title":"Members"}},"type":"object","required":["name","faculty_id","id"],"title":"DepartmentOut"},"DepartmentUpdate":{"properties":{"name":{"anyOf":[{"type":"string"},{"type":"null"}],"title":"Name"},"faculty_id":{"anyOf":[{"type":"string","format":"uuid"},{"type":"null"}],"title":"Faculty Id"}},"type":"object","title":"DepartmentUpdate"},"Enrollment":{"properties":{"user_id":{"type":"string","format":"uuid","title":"User Id"},"course_id":{"type":"string","format":"uuid","title":"Course Id"},"id":{"anyOf":[{"type":"string","format":"uuid"},{"type":"null"}],"title":"Id"},"enrolled_at":{"type":"string","format":"date-time","title":"Enrolled At"}},"type":"object","required":["user_id","course_id"],"title":"Enrollment"},"EnrollmentOut":{"properties":{"user_id":{"type":"string","format":"uuid","title":"User Id"},"course_id":{"type":"string","format":"uuid","title":"Course Id"},"id":{"type":"string","format":"uuid","title":"Id"},"course":{"anyOf":[{"$ref":"#/components/schemas/Course"},{"type":"null"}]}},"type":"object","required":["user_id","course_id","id","course"],"title":"EnrollmentOut"},"FaceEncodingOut":{"properties":{"id":{"type":"string","format":"uuid","title":"Id"},"created_at":{"type":"string","format":"date-time","title":"Created At"}},"type":"object","required":["id","created_at"],"title":"FaceEncodingOut"},"FacultyIn":{"properties":{"name":{"type":"string","title":"Name"},"code":{"type":"string","title":"Code"},"description":{"anyOf":[{"type":"string"},{"type":"null"}],"title":"Description"}},"type":"object","required":["name","code"],"title":"FacultyIn"},"FacultyOut":{"properties":{"name":{"type":"string","title":"Name"},"code":{"type":"string","title":"Code"},"description":{"anyOf":[{"type":"string"},{"type":"null"}],"title":"Description"},"id":{"type":"string","format":"uuid","title":"Id"}},"type":"object","required":["name","code","id"],"title":"FacultyOut"},"FacultyOutWithDepartment":{"properties":{"name":{"type":"string","title":"Name"},"code":{"type":"string","title":"Code"},"description":{"anyOf":[{"type":"string"},{"type":"null"}],"title":"Description"},"id":{"type":"string","format":"uuid","title":"Id"},"departments":{"anyOf":[{"items":{"$ref":"#/components/schemas/Department"},"type":"array"},{"type":"null"}],"title":"Departments"}},"type":"object","required":["name","code","id"],"title":"FacultyOutWithDepartment"},"FacultyUpdate":{"properties":{"name":{"anyOf":[{"type":"string"},{"type":"null"}],"title":"Name"},"code":{"type":"string","title":"Code"},"description":{"anyOf":[{"type":"string"},{"type":"null"}],"title":"Description"}},"type":"object","required":["code"],"title":"FacultyUpdate"},"HTTPValidationError":{"properties":{"detail":{"items":{"$ref":"#/components/schemas/ValidationError"},"type":"array","title":"Detail"}},"type":"object","title":"HTTPValidationError"},"Msg":{"properties":{"msg":{"type":"string","title":"Msg"}},"type":"object","required":["msg"],"title":"Msg"},"Token":{"properties":{"access_token":{"type":"string","title":"Access Token"},"token_type":{"type":"string","title":"Token Type"}},"type":"object","required":["access_token","token_type"],"title":"Token"},"User":{"properties":{"email":{"type":"string","title":"Email"},"full_name":{"anyOf":[{"type":"string"},{"type":"null"}],"title":"Full Name"},"level":{"anyOf":[{"type":"string"},{"type":"null"}],"title":"Level"},"age":{"anyOf":[{"type":"integer"},{"type":"null"}],"title":"Age"},"phone_number":{"anyOf":[{"type":"string"},{"type":"null"}],"title":"Phone Number"},"home_address":{"anyOf":[{"type":"string"},{"type":"null"}],"title":"Home Address"},"department_id":{"anyOf":[{"type":"string","format":"uuid"},{"type":"null"}],"title":"Department Id"},"id":{"anyOf":[{"type":"string","format":"uuid"},{"type":"null"}],"title":"Id"},"disabled":{"type":"boolean","title":"Disabled","default":false},"role":{"$ref":"#/components/schemas/UserRoles"},"created_at":{"type":"string","format":"date-time","title":"Created At"},"hashed_password":{"type":"string","title":"Hashed Password"}},"type":"object","required":["email","role","hashed_password"],"title":"User"},"UserAdminUpdate":{"properties":{"disabled":{"anyOf":[{"type":"boolean"},{"type":"null"}],"title":"Disabled"},"role":{"anyOf":[{"$ref":"#/components/schemas/UserRoles"},{"type":"null"}]}},"additionalProperties":false,"type":"object","title":"UserAdminUpdate"},"UserIn":{"properties":{"email":{"type":"string","title":"Email"},"full_name":{"anyOf":[{"type":"string"},{"type":"null"}],"title":"Full Name"},"level":{"anyOf":[{"type":"string"},{"type":"null"}],"title":"Level"},"age":{"anyOf":[{"type":"integer"},{"type":"null"}],"title":"Age"},"phone_number":{"anyOf":[{"type":"string"},{"type":"null"}],"title":"Phone Number"},"home_address":{"anyOf":[{"type":"string"},{"type":"null"}],"title":"Home Address"},"department_id":{"anyOf":[{"type":"string","format":"uuid"},{"type":"null"}],"title":"Department Id"},"password":{"type":"string","title":"Password"}},"type":"object","required":["email","password"],"title":"UserIn"},"UserOut":{"properties":{"email":{"type":"string","title":"Email"},"full_name":{"anyOf":[{"type":"string"},{"type":"null"}],"title":"Full Name"},"level":{"anyOf":[{"type":"string"},{"type":"null"}],"title":"Level"},"age":{"anyOf":[{"type":"integer"},{"type":"null"}],"title":"Age"},"phone_number":{"anyOf":[{"type":"string"},{"type":"null"}],"title":"Phone Number"},"home_address":{"anyOf":[{"type":"string"},{"type":"null"}],"title":"Home Address"},"department_id":{"anyOf":[{"type":"string","format":"uuid"},{"type":"null"}],"title":"Department Id"},"id":{"type":"string","format":"uuid","title":"Id"},"enrollments":{"anyOf":[{"items":{"$ref":"#/components/schemas/Enrollment"},"type":"array"},{"type":"null"}],"title":"Enrollments"}},"type":"object","required":["email","id","enrollments"],"title":"UserOut"},"UserRoles":{"type":"string","enum":["STUDENT","ADMIN","LECTURER"],"title":"UserRoles"},"UserUpdate":{"properties":{"full_name":{"anyOf":[{"type":"string"},{"type":"null"}],"title":"Full Name"},"age":{"anyOf":[{"type":"integer"},{"type":"null"}],"title":"Age"},"email":{"anyOf":[{"type":"string"},{"type":"null"}],"title":"Email"},"phone_number":{"anyOf":[{"type":"string"},{"type":"null"}],"title":"Phone Number"},"home_address":{"anyOf":[{"type":"string"},{"type":"null"}],"title":"Home Address"},"password":{"anyOf":[{"type":"string"},{"type":"null"}],"title":"Password"}},"additionalProperties":false,"type":"object","title":"UserUpdate"},"ValidationError":{"properties":{"loc":{"items":{"anyOf":[{"type":"string"},{"type":"integer"}]},"type":"array","title":"Location"},"msg":{"type":"string","title":"Message"},"type":{"type":"string","title":"Error Type"}},"type":"object","required":["loc","msg","type"],"title":"ValidationError"}},"securitySchemes":{"OAuth2PasswordBearer":{"type":"oauth2","flows":{"password":{"scopes":{},"tokenUrl":"auth/token"}}}}}}